chore: Update Unity and GVR Versions Fix: Shaders are no longer unlit
Unity version has been upgraded from 2017.4.4 to 2017.4.15 which is the
latest LTS version.
Google VR sdk has been upgraded to 1.170.0 which is the latest version.

The shaders on the apartment walls, doors and windows have been changed
from mobile/unlit to standard to be consistent with the course content
at this point. (The unlit shaders were confusing students with a lack of
lighting on the walls.)